Changed the definition of length encoding
Now we use two bytes (parametricable) to represent the length of
elements instead of byte size.

+++ share/ct-fuzz/include/ct-fuzz.h
@@ -1,11 +1,12 @@
 #ifndef CT_FUZZ_H
 #define CT_FUZZ_H
 #include <stdbool.h>
+#include <stdint.h>
 #include "ct-fuzz-observation.h"
 
 void PREFIX(public_in)();
 void PREFIX(assume)(bool, char* msg);
-size_t PREFIX(array_len)(void*);
+uint16_t PREFIX(array_len)(void*);
 
 #define CT_FUZZ_SPEC(F) void __ct_fuzz_spec_ ## F
 #define CT_FUZZ_ASSUME(cond) PREFIX(assume)(cond, "assumption violation")
